refactor(events) : put EventCard.kt into ui directory for consitency purposes
